- Update USR's code
git-svn-id: https://svn.code.sf.net/p/speed-dreams/code/trunk@6993 30fe4595-0a0c-4342-8851-515496e4dcbd

Former-commit-id: 288862ea9cd5d918f909ba96cc9099210727a94a
Former-commit-id: 95c7c44be9d7db27201ac84f2d2b54750627ae2c

+++ src/drivers/usr/src/driver.cpp
@@ -112,7 +112,7 @@ void Driver::InitTrack(const tTrack* Track, void* carHandle, void** carParmHandl
     mPitEntryMargin = param.getNum("private", "pitentrymargin");
     mPitEntrySpeed = param.getNum("private", "pitentryspeed");
     mPitExitSpeed = param.getNum("private", "pitexitspeed");
-    mPitTest = param.getNum("private", "pittest");
+    mPitTest = (int)param.getNum("private", "pittest");
     mSegLen = param.getNum("private", "seglen");
     mClothFactor = param.getNum("private", "clothoidfactor");
     mVMaxK = param.getNum("private", "vmaxk");
